- move fetch data from firebase when login logic from App.tsx to Header.tsx - prevent render when typing in search input at header with using useRef - unmount redirect to home page function on logo(if user don't log yet, they click the log the page will redirect to the same page, so this redirect is unnecessary - remove redundant state with declare a new variable instead(variable with global state in it so if global state change, this variable redeclared a new variable - prevent header jump into default status when is loged(when refresh the page header shouldn't render unlogged status first then turn into logged status
